feat: Integrate with Cloud func for app update

+++ data/common/src/main/java/com/popstack/mvoter2015/data/common/appupdate/AppUpdate.kt
@@ -1,7 +1,7 @@
 package com.popstack.mvoter2015.data.common.appupdate
 
 data class AppUpdate(
-  val latestVersionCode: Int,
+  val latestVersionCode: Long,
   val requireForcedUpdate: Boolean,
   val playStoreLink: String,
   val selfHostedLink: String
